misc/gpu/requirements.txt
# optional, unless you want to run make_barcode.py
python-barcode==0.15.1
